separate out exported function into separate file
