+++ src/lambdas/RedriveSNSDQLEvents/writeToFirehose.ts
@@ -1,10 +1,10 @@
-import { PutRecordBatchCommandOutput } from '@aws-sdk/client-firehose'
 import { firehosePutRecordBatch } from '../../sharedServices/firehose/firehosePutRecordBatch'
 import { logger } from '../../sharedServices/logger'
 import { AuditEvent } from '../../types/auditEvent'
 import { auditEventsToFirehoseRecords } from '../../utils/helpers/firehose/auditEventsToFirehoseRecords'
 import { getEnv } from '../../utils/helpers/getEnv'
 import { ProcessingResult } from './helper'
+import { parseFirehoseResponse } from './parseFirehoseResponse'
 
 export type FirehoseProcessingResult = {
   successfullProcessingResults: ProcessingResult[]
@@ -33,50 +33,5 @@ export const writeToFirehose = async (
       failedProcessingResults: successfullyParsedRecords
     }
     return errorOutput
-  }
-}
-
-export const parseFirehoseResponse = (
-  firehoseResponse: PutRecordBatchCommandOutput,
-  processingResults: ProcessingResult[]
-): FirehoseProcessingResult => {
-  const successMessage = 'SucceededToWriteToFirehose'
-
-  if (firehoseResponse.FailedPutCount && firehoseResponse.FailedPutCount > 0) {
-    logger.warn('Some audit events failed to reingest')
-    const successfullProcessingResults: ProcessingResult[] = []
-    const failedProcessingResults: ProcessingResult[] = []
-
-    firehoseResponse.RequestResponses?.forEach((response, index) => {
-      if (response.ErrorCode && processingResults[index]) {
-        processingResults[index] = {
-          ...processingResults[index],
-          statusReason: 'FailedToWriteToFirehose',
-          failed: true
-        }
-        failedProcessingResults.push(processingResults[index])
-      } else {
-        processingResults[index] = {
-          ...processingResults[index],
-          statusReason: successMessage,
-          failed: false
-        }
-        successfullProcessingResults.push(processingResults[index])
-      }
-    })
-    return {
-      successfullProcessingResults,
-      failedProcessingResults
-    }
-  } else {
-    return {
-      successfullProcessingResults: processingResults.map((element) => {
-        return {
-          ...element,
-          statusReason: successMessage
-        }
-      }),
-      failedProcessingResults: []
-    }
   }
 }
